fix: Allow player to move after a collision
Check collision using the next position instead of current position when checking for collisions. This way once a collision has happened, the player can still move as their bounding box doesn't end up inside the collision rectangle.

There is still a bug related to this when using mouse movement. As this isn't true collision, the player can walk through other characters if the next position is beyond the other characters.